release tag is updated...

+++ tags/release/sdl/dataitem.cpp
@@ -8,16 +8,18 @@
 
 #include "dataitem.h"
 #include "estrbuffer.h"
+#include "exparse/exptree.h"
 
   static olxstr DoubleQuoteCode("&2E;");
 
 UseEsdlNamespace()
+using namespace exparse::parser_util;
 
 //..............................................................................
 TDataItem::TDataItem(TDataItem* Prnt, const olxstr& nm, const olxstr& val) : Name(nm), Value(val)  {
   if( Prnt != NULL )  {
     Level = Prnt->GetLevel()+1;
-    Index = Prnt->ItemCount();
+    Index = (int)Prnt->ItemCount();
   }
   else  {
     Level = 0;
@@ -27,13 +29,9 @@ TDataItem::TDataItem(TDataItem* Prnt, const olxstr& nm, const olxstr& val) : Nam
   Data = NULL;
 }
 //..............................................................................
-TDataItem::~TDataItem()  {
-  Clear();
-}
-//..............................................................................
 void TDataItem::Clear()  {
   Fields.Clear();
-  for( int i=0; i < Items.Count(); i++ )  {
+  for( size_t i=0; i < Items.Count(); i++ )  {
     if( Items.GetObject(i)->GetRefCount() < 1 ) 
       delete Items.GetObject(i);
     else  {
@@ -43,10 +41,6 @@ void TDataItem::Clear()  {
     }
   }
   Items.Clear();
-}
-//..............................................................................
-bool TDataItem::ItemExists(const olxstr &Name)  {
-  return (Items.IndexOf(Name) == -1) ? false : true;
 }
 //..............................................................................
 TDataItem& TDataItem::Root()  {
@@ -68,8 +62,8 @@ TEStrBuffer& TDataItem::writeFullName(TEStrBuffer& bf) const {
       tmp.Add(P);
       P = P->GetParent();
     }
-    for( int i=tmp.Count()-1; i >= 0; i-- )
-      bf << tmp[i]->GetName() << '.';
+    for( size_t i=tmp.Count(); i > 0; i-- )
+      bf << tmp[i-1]->GetName() << '.';
     bf << Name;
   }
   return bf;
@@ -91,10 +85,11 @@ olxstr TDataItem::GetFullName()  {
 TDataItem *TDataItem::DotItem(const olxstr &DotName, TStrList* Log)  {
   TDataItem* root = &Root();
   TStrList SL(DotName, '.');
-  for( int i=0; i < SL.Count(); i++ )  {
+  for( size_t i=0; i < SL.Count(); i++ )  {
     root = root->FindItem( SL[i] );
     if( root == NULL )  {
-      if( Log != NULL )  Log->Add(olxstr("Unresolved reference: ") << DotName);
+      if( Log != NULL )  
+        Log->Add(olxstr("Unresolved reference: ") << DotName);
       break;
     }
   }
@@ -105,7 +100,7 @@ olxstr* TDataItem::DotField(const olxstr& DotName, olxstr& RefFieldName)  {
   TDataItem* root = &Root(), *PrevItem;
   TStrList SL(DotName, '.');
   olxstr *Str=NULL;
-  for( int i=0; i < SL.Count(); i++ )  {
+  for( size_t i=0; i < SL.Count(); i++ )  {
     PrevItem = root;
     root = root->FindItem( SL[i] );
     if( root == NULL )  {
@@ -120,7 +115,7 @@ olxstr* TDataItem::DotField(const olxstr& DotName, olxstr& RefFieldName)  {
 TDataItem *TDataItem::GetAnyItem(const olxstr& Name) const {
   TDataItem *DI = Items.FindObject(Name);
   if( DI == NULL )  {
-    for( int i=0; i < ItemCount(); i++ )  {
+    for( size_t i=0; i < ItemCount(); i++ )  {
       DI = GetItem(i).GetAnyItem(Name);
       if( DI != NULL )  return DI;
     }
@@ -131,7 +126,7 @@ TDataItem *TDataItem::GetAnyItem(const olxstr& Name) const {
 TDataItem *TDataItem::GetAnyItemCI(const olxstr &Name) const {
   TDataItem *DI = Items.FindObjecti(Name);
   if( DI == NULL )  {
-    for( int i=0; i < ItemCount(); i++ )  {
+    for( size_t i=0; i < ItemCount(); i++ )  {
       DI = GetItem(i).GetAnyItemCI(Name);
       if( DI != NULL )  return DI;
     }
@@ -139,13 +134,25 @@ TDataItem *TDataItem::GetAnyItemCI(const olxstr &Name) const {
   return DI;
 }
 //..............................................................................
-void TDataItem::AddContent(TDataItem& DI)  {
-  TDataItem *di;
-  for( int i=0; i < DI.ItemCount(); i++ )  {
-    di = FindItem( DI.GetItem(i).GetName() );
+void TDataItem::AddContent(TDataItem& DI, bool extend)  {
+  if( extend )  {
+    for( size_t i=0; i < DI.ItemCount(); i++ )  {
+      TDataItem *di = FindItem( DI.GetItem(i).GetName() );
+      if( di != NULL )  {
+        for( size_t j=0; j < DI.GetItem(i).ItemCount(); j++ )
+          di->AddItem(DI.GetItem(i).GetItem(j)).SetParent(di);
+      }
+      else
+        AddItem(DI.GetItem(i)).SetParent(this);
+    }
+  }
+  else  {
+    for( size_t i=0; i < DI.ItemCount(); i++ )  {
+      TDataItem *di = FindItem( DI.GetItem(i).GetName() );
     if( di != NULL )
       DeleteItem(di);
-    AddItem( DI.GetItem(i) );
+      AddItem(DI.GetItem(i)).SetParent(this);
+  }
   }
 }
 //..............................................................................
@@ -155,8 +162,8 @@ void TDataItem::DeleteItem(TDataItem *Item)  {
       Item->GetParent()->DeleteItem(Item);
     return;
   }
-  int index = Items.IndexOfObject(Item);
-  if( index >= 0 )  Items.Delete(index);
+  size_t index = Items.IndexOfObject(Item);
+  if( index != InvalidIndex )  Items.Delete(index);
   if( Item->GetRefCount() < 1 )
     delete Item;
   else  {
@@ -189,89 +196,76 @@ TDataItem& TDataItem::AddItem(const olxstr& name, const olxstr& val)  {
   return *DI;
 }
 //..............................................................................
-int TDataItem::LoadFromString( int start, olxstr &Data, TStrList* Log)  {
-  int sl = Data.Length();
-  TDataItem *DI;
-  Value = EmptyString;
-  olxstr ItemName, FieldName, FieldValue, *RefField, RefFieldName;
-  olxch CloseChar;
-  for( int i=start; i < sl; i++ )  {
-    if( Data.CharAt(i) == '<' )  {
-      ItemName = EmptyString;
-      while( i < sl )  {
-        i++;
-        if( Data.CharAt(i) == ' ' || Data.CharAt(i) == '\"' )  break;
-        ItemName << Data.CharAt(i);
+size_t TDataItem::LoadFromString(size_t start, const olxstr &Data, TStrList* Log)  {
+  const size_t sl = Data.Length();
+  for( size_t i=start; i < sl; i++ )  {
+    olxch ch = Data.CharAt(i);
+    if( ch == '<' )  {
+      const size_t name_start_i = i+1;
+      while( ++i < sl )  {
+        ch = Data.CharAt(i);
+        if( ch == ' ' || ch == '<' || ch == '>' || ch == '\"' )  break;
       }
+      olxstr ItemName = Data.SubString(name_start_i, i-name_start_i);
       if( ItemName.IsEmpty() && Log != NULL )
         Log->Add((this->GetName() + ':') << " empty item name!");
-      DI = NULL;
+      TDataItem* DI = NULL;
       try {  DI = &AddItem(ItemName);  }
       catch(...) {
         if( Log != NULL )
           Log->Add((this->GetName() + ':') << " cannot add item");
       }
       if( DI != NULL )  {
-//        BasicApp->Log->CriticalInfo(((olxstr("Loading: ") += this->Name()) += '.') += ItemName);
         i = DI->LoadFromString(i, Data, Log);
         continue;
       }
     }
 
-    if( Data.CharAt(i) == '>' )  return i;
-    if( Data.CharAt(i) == '/' )  continue;
-    if( Data.CharAt(i) == '\\' ) continue;
-    if( Data.CharAt(i) != ' ')   {
-      FieldName = EmptyString;
-      FieldValue = EmptyString;
-      if( Data.CharAt(i) == '\'' || Data.CharAt(i) == '\"' )  {  // item value
-        CloseChar = Data.CharAt(i);  // find closing character: space ' or "
-        i++;
-        while( i < sl )  {
-          if( (Data.CharAt(i) == CloseChar) && (Data.CharAt(i-1) != '^') ){  i++; break; }
-          Value << Data.CharAt(i);
-          i++;
-        }
+    if( ch == '>' )  return i;
+    if( ch == '/' )  continue;
+    if( ch == '\\' ) continue;
+    if( ch != ' ')   {
+      olxstr FieldName = EmptyString;
+      olxstr FieldValue = EmptyString;
+      if( is_quote(ch) )  {  // item value
+        parse_escaped_string(Data, Value, i);
         if( i < sl && (Data.CharAt(i) == '>') )  {  return i;  }
         continue;
       }
       if( (i+1) >= sl ) return sl+1;
+      const size_t fn_start = i;
       while( i < sl )  {  // extract field name
-        if( (Data.CharAt(i) == ' ') || (Data.CharAt(i) == '=') ||
-            (Data.CharAt(i) == '>') || (Data.CharAt(i) == '\"') ||
-            (Data.CharAt(i) == '\''))  break;
-        FieldName << Data.CharAt(i);
+        ch = Data.CharAt(i);
+        if( ch == ' ' || ch == '=' || ch == '>' || 
+            ch == '\"' || ch == '\'' || ch == '<')  break;
         i++;
       }
+      FieldName = Data.SubString(fn_start, i - fn_start);
       if( (i+1) >= sl ) return sl+1;
-      while( i < sl )  {  // skip spaces
-        if( Data.CharAt(i) != ' ' )  break;
-        i++;
-      }
+      if( Data.CharAt(i) == ' ' )  i++;  // skip space
       if( (i+1) >= sl ) return sl+1;
       if( Data.CharAt(i) == '=' )  {  // extract field value
         i++;
-        while( i < sl )  {  // skip spaces
-          if( Data.CharAt(i) != ' ' )  break;
-          i++;
-        }
+        if( Data.CharAt(i) == ' '  )  i++;  // skip space
         if( (i+1) >= sl ) return sl+1;
-        CloseChar = ' ';  // find closing character: space ' or "
-        if( Data.CharAt(i) == '\'' ){  CloseChar = '\''; i++; }
-        if( Data.CharAt(i) == '\"' ){  CloseChar = '\"'; i++; }
+        olxch CloseChar = ' ';  // find closing character: space ' or "
+        if( Data.CharAt(i) == '\'' )  {  CloseChar = '\''; i++; }
+        else if( Data.CharAt(i) == '\"' )  {  CloseChar = '\"'; i++; }
+        const size_t fv_start = i;
         while( i < sl )  {
-          if( (Data.CharAt(i) == CloseChar) && (Data[i-1] != '^') ){  i++; break; }
-//          if( (Data.CharAt(i) == CloseChar) || (Data.CharAt(i) == '>') ){  i++; break; }
-          FieldValue << Data.CharAt(i);
+          if( Data.CharAt(i) == CloseChar || (CloseChar == ' ' && Data.CharAt(i) == '>') )  break;
           i++;
         }
+        FieldValue = Data.SubString(fv_start, i-fv_start);
       }
       else i--; // the spaces can separate just two field names
-      if( FieldName.IndexOf('.') >= 0 )  {  // a reference to an item
-        DI = DotItem(FieldName, Log);
-        if( DI != NULL )  AddItem(*DI); // unresolved so far if !DI
-        else  {
-          RefField = DotField(FieldName, RefFieldName);
+      if( FieldName.IndexOf('.') != InvalidIndex )  {  // a reference to an item
+        TDataItem* DI = DotItem(FieldName, Log);
+        if( DI != NULL )  
+          AddItem(*DI);
+        else  {  // unresolved so far if !DI
+          olxstr RefFieldName;
+          olxstr *RefField = DotField(FieldName, RefFieldName);
           if( RefField == NULL )
             _AddField(FieldName, FieldValue); 
           else
@@ -289,25 +283,25 @@ int TDataItem::LoadFromString( int start, olxstr &Data, TStrList* Log)  {
 }
 //..............................................................................
 void TDataItem::ResolveFields(TStrList* Log)  {  // resolves referenced fields
-  olxstr *RefFieldValue, RefFieldName;
-  TDataItem *DI;
-  for( int i=0; i < FieldCount(); i++ )  {
+  for( size_t i=0; i < FieldCount(); i++ )  {
     const olxstr& Tmp = Fields[i];
-    if( Tmp.IndexOf('.') >= 0 )  {  // a reference to an item
-      DI = DotItem(Tmp, Log);
+    if( Tmp.IndexOf('.') != InvalidIndex )  {  // a reference to an item
+      TDataItem *DI = DotItem(Tmp, Log);
       if( DI != NULL )  {
         AddItem(*DI);
-        if( Log != NULL )  Log->Add( olxstr("Resolved: ") << Tmp);
+        if( Log != NULL )
+          Log->Add( olxstr("Resolved: ") << Tmp);
         Fields.Delete(i);  // might be very slow !!!
         i--;
       }
       else  {
-        RefFieldValue = DotField(Tmp, RefFieldName);
+        olxstr RefFieldName(EmptyString);
+        olxstr *RefFieldValue = DotField(Tmp, RefFieldName);
         if( RefFieldValue == NULL )  {
-          if( Log != NULL )  Log->Add(olxstr("UNRESOLVED: ") << Tmp);
+          if( Log != NULL )
+            Log->Add(olxstr("UNRESOLVED: ") << Tmp);
         }
-        else
-        {
+        else  {
           if( Log)  Log->Add(olxstr("Resolved field: ") << Tmp);
           Fields[i] = RefFieldName;
           Fields.GetObject(i) = *RefFieldValue;
@@ -315,14 +309,13 @@ void TDataItem::ResolveFields(TStrList* Log)  {  // resolves referenced fields
       }
     }
   }
-  for( int i=0; i < ItemCount(); i++ )  {
+  for( size_t i=0; i < ItemCount(); i++ )  {
     if( GetItem(i).GetParent() == this )
       GetItem(i).ResolveFields(Log);
   }
 }
 //..............................................................................
 void TDataItem::SaveToStrBuffer(TEStrBuffer &Data) const {
-  int fc, ic;
   bool itemsadded = false;
   if( GetParent() != NULL )  {
     if( Data.Length() != 0 )
@@ -333,19 +326,19 @@ void TDataItem::SaveToStrBuffer(TEStrBuffer &Data) const {
       Data << '\"' << Value << "\" ";
     }
   }
-  fc = FieldCount();
-  ic = ItemCount();
-  for( int i=0; i < fc; i++ )  {
+  const size_t fc = FieldCount();
+  for( size_t i=0; i < fc; i++ )  {
     Data << Fields.GetString(i) << '=';
     Data << '\"' << CodeString(Fields.GetObject(i)) << '\"' << ' ';
   }
-  for( int i=0; i < ic; i++ )  {
-    if( GetItem(i).GetParent() != this )  {  // dot operator
+  const size_t ic = ItemCount();
+  for( size_t i=0; i < ic; i++ )  {
+    if( GetItem(i).GetParent() != this && GetItem(i).GetParent() != NULL )  {  // dot operator
       GetItem(i).writeFullName(Data) << ' ';
     }
   }
-  for( int i=0; i < ic; i++ )  {
-    if( GetItem(i).GetParent() == this )  {
+  for( size_t i=0; i < ic; i++ )  {
+    if( GetItem(i).GetParent() == this || GetItem(i).GetParent() == NULL )  {
       GetItem(i).SaveToStrBuffer(Data);
       if( !itemsadded ) itemsadded = true;
     }
@@ -363,21 +356,21 @@ void TDataItem::SaveToStrBuffer(TEStrBuffer &Data) const {
 }
 //..............................................................................
 void TDataItem::FindSimilarItems(const olxstr &StartsFrom, TPtrList<TDataItem>& List)  {
-  for(int i=0; i < ItemCount(); i++ )  {
+  for( size_t i=0; i < ItemCount(); i++ )  {
     if( Items[i].StartsFromi(StartsFrom) )
       List.Add(Items.GetObject(i));
   }
 }
 //..............................................................................
 olxstr TDataItem::CodeString(const olxstr& Str) const  {
-  if( Str.IndexOf('\"') == -1 )  return Str;
+  if( Str.IndexOf('\"') == InvalidIndex )  return Str;
   olxstr rv(Str);
   rv.Replace('\"', DoubleQuoteCode);
   return rv;
 }
 //..............................................................................
 olxstr TDataItem::DecodeString(const olxstr& Str) const  {
-  if( Str.IndexOf(DoubleQuoteCode) == -1 )  return Str;
+  if( Str.IndexOf(DoubleQuoteCode) == InvalidIndex )  return Str;
   olxstr Res(Str);
   Res.Replace(DoubleQuoteCode, '\"');
   return Res;
@@ -391,7 +384,7 @@ void TDataItem::Sort()  {
 //..............................................................................
 void TDataItem::Validate(TStrList& Log)  {
   olxstr Tmp;
-  for( int i=0; i < ItemCount(); i++ )  {
+  for( size_t i=0; i < ItemCount(); i++ )  {
     if( GetItem(i).GetParent() != NULL && (GetItem(i).GetParent() != this) )  {
       Tmp = "Reffered item from ";
       Tmp << this->GetFullName() << "->" << GetItem(i).GetFullName();
